Create component to display reviews
This component was a challenge to complete. The component starts with the specified amount of empty stars. If the user hovers over a star it shows how many stars are going to be filled when the user clicks. If the user clicks, the rating is saved to the state and the amount of stars selected become full, if not, the stars stay as they were before the user hovered over.