Fix useEffect and add components
Added a condition to the effect so that it only runs when the query is not empty or the query length is greater than 2. Fixed an issue where the error was not reseted and this prevented the movies to display correctly. Add the movie detail component to display the information of the selected movie